Type changes
Provides the division with a fractional part in a result. 9/2=4.5 (not 4)

+++ calculator/calcmodel.h
@@ -13,13 +13,13 @@ public:
 
 private:
     int _state;
-    int _currentValue;
-    int _outputValue;
-    int _operand1;
+    float _currentValue;
+    float _outputValue;
+    float _operand1;
 
     QString _operator;
     QString _display;
-    int calculator(int value1, int value2, QString inputOp);
+    float calculator(float value1, float value2, QString inputOp);
 };
 
 #endif // CALCMODEL_H
